Added user authentication using json web token.

+++ controller/contactControllers.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.body.user.id });
   res.status(200).send(contacts);
 });
 
@@ -25,6 +25,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id
   });
   res.status(201).send(contact);
 });
@@ -36,6 +37,12 @@ const updateContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found!");
+  }
+
+  // if a differnet user is trying to update a contact
+  if(req.user.id != contact.user_id) {
+    res.status(403);
+    throw new Error("Don't have access to update contact");
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(id, body, {
@@ -51,6 +58,13 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found!");
   }
+
+  // if a differnet user is trying to delete a contact
+  if(req.user.id !== contact.user_id) {
+    res.status(403);
+    throw new Error("Don't have access to delete contact");
+  }
+
   const deletedContact = await Contact.findByIdAndDelete(id);
   if (!deletedContact) {
     res.status(404);
